chore: organizes CMS sections
