+++ packages/core/src/components/common/Footer/section.module.scss
@@ -7,5 +7,6 @@
   @import "@faststore/ui/src/components/atoms/List/styles.scss";
   @import "@faststore/ui/src/components/molecules/Accordion/styles.scss";
   @import "@faststore/ui/src/components/organisms/PaymentMethods/styles.scss";
+  @import "@faststore/ui/src/components/organisms/Incentives/styles.scss";
   @import "@faststore/ui/src/components/organisms/Footer/styles.scss";
 }
