--- packages/core/src/components/sections/Incentives/IncentivesFooter.tsx
@@ -1,8 +1,0 @@
-import Incentives from './Incentives'
-import { incentivesMockFooter as IncentivesMock } from './incentivesMock'
-
-function IncentivesFooter() {
-  return <Incentives incentives={IncentivesMock} />
-}
-
-export default IncentivesFooter
