--- packages/core/src/components/sections/Incentives/incentivesMock.ts
@@ -1,55 +1,0 @@
-export const incentivesMockHeader = [
-  {
-    icon: 'Truck',
-    title: 'Buy online',
-    firstLineText: 'Get Free Shipping',
-  },
-  {
-    icon: 'Calendar',
-    title: 'Free return',
-    firstLineText: '30 days to return',
-  },
-  {
-    icon: 'Gift',
-    title: 'Gift cards',
-    firstLineText: '$20 / $30 / $50',
-  },
-  {
-    icon: 'Storefront',
-    title: 'Physical Stores',
-    firstLineText: '+40 Stores in Brazil',
-  },
-  {
-    icon: 'ShieldCheck',
-    title: 'Buy online',
-    firstLineText: 'Get Free Shipping',
-  },
-]
-
-export const incentivesMockFooter = [
-  {
-    icon: 'ShieldCheck',
-    firstLineText: 'Trusted',
-    secondLineText: 'by Safecon',
-  },
-  {
-    icon: 'Medal',
-    firstLineText: 'Quality',
-    secondLineText: 'Products',
-  },
-  {
-    icon: 'CircleWavyCheck',
-    firstLineText: '3-year',
-    secondLineText: 'Guarantee',
-  },
-  {
-    icon: 'Storefront',
-    firstLineText: 'Pickup',
-    secondLineText: 'Options',
-  },
-  {
-    icon: 'Truck',
-    firstLineText: 'Free',
-    secondLineText: 'Shippping',
-  },
-]
