+++ packages/core/src/components/sections/IncentivesHeader/index.ts
@@ -1,0 +1,1 @@
+export { default } from './IncentivesHeader'
