+++ packages/core/src/components/ui/Incentives/incentivesMock.ts
@@ -1,0 +1,27 @@
+export const incentivesMockFooter = [
+  {
+    icon: 'ShieldCheck',
+    firstLineText: 'Trusted',
+    secondLineText: 'by Safecon',
+  },
+  {
+    icon: 'Medal',
+    firstLineText: 'Quality',
+    secondLineText: 'Products',
+  },
+  {
+    icon: 'CircleWavyCheck',
+    firstLineText: '3-year',
+    secondLineText: 'Guarantee',
+  },
+  {
+    icon: 'Storefront',
+    firstLineText: 'Pickup',
+    secondLineText: 'Options',
+  },
+  {
+    icon: 'Truck',
+    firstLineText: 'Free',
+    secondLineText: 'Shippping',
+  },
+]
