+++ packages/core/src/pages/index.tsx
@@ -6,7 +6,7 @@ import type { ComponentType } from 'react'
 import RenderPageSections from 'src/components/cms/RenderPageSections'
 import BannerText from 'src/components/sections/BannerText'
 import Hero from 'src/components/sections/Hero'
-import IncentivesHeader from 'src/components/sections/Incentives/IncentivesHeader'
+import IncentivesHeader from 'src/components/sections/IncentivesHeader/IncentivesHeader'
 import Newsletter from 'src/components/sections/Newsletter'
 import ProductShelf from 'src/components/sections/ProductShelf'
 import ProductTiles from 'src/components/sections/ProductTiles'
